interop: add genesis activation support and align ETHLiquidity funding
- Add SuperchainETHBridge and ETHLiquidity to L2Genesis script so both
  contracts are deployed when interop activates at genesis (offset 0)
- Align ETHLiquidity funding to uint128 max across genesis and upgrade
  paths (constrained by deposit tx mint field)
- Add TestInteropGenesisActivation to verify all interop contracts are
  deployed and ETHLiquidity is funded at genesis

+++ packages/contracts-bedrock/scripts/L2Genesis.s.sol
@@ -550,7 +550,7 @@ contract L2Genesis is Script {
     ///         This contract has no initializer.
     function setETHLiquidity() internal {
         _setImplementationCode(Predeploys.ETH_LIQUIDITY);
-        vm.deal(Predeploys.ETH_LIQUIDITY, type(uint248).max);
+        vm.deal(Predeploys.ETH_LIQUIDITY, type(uint128).max);
     }
 
     /// @notice This predeploy is following the safety invariant #1.
